renamed AuthManagement to SessionManagement
